+++ src/app/shared/main-layout/header/header.component.html
@@ -1,10 +1,10 @@
 <header>
     <app-logo></app-logo>
-    <div class="header__user-details" *ngIf="$isAuthenticated | async">
+    <div class="header__user-details" *ngIf="isAuthenticated$ | async">
         <div class="user-details__login">
             <img alt="User Icon" class="user-details__user-icon" src="../../../../assets/img/ic-user.svg"/>
             <span>
-                {{ this.$user | async }}
+                {{ this.user$ | async }}
             </span>
         </div>
