+++ src/app/shared/pipes/index.ts
@@ -1,3 +1,3 @@
-export { DurationPipe } from "./duration.pipe";
-export { OrderByPipe } from "./order-by.pipe";
-export { FilterPipe } from "./filter.pipe";
+export { DurationPipe } from "./duration/duration.pipe";
+export { OrderByPipe } from "./order-by/order-by.pipe";
+export { FilterPipe } from "./filter/filter.pipe";
