+++ src/app/shared/pipes/order-by/order-by.pipe.ts
@@ -1,5 +1,5 @@
 import { Pipe, PipeTransform } from "@angular/core";
-import { ICourse } from "../../core/index";
+import { ICourse } from "../../../core/index";
 
 @Pipe({
     name: "orderBy"
